+++ test/tree-sitter/testdata/c_sharp/StepDefinitions.cs
@@ -1,0 +1,12 @@
+using TechTalk.SpecFlow;
+
+namespace StepDefinitions
+{
+    [Binding]
+    public class BowlingSteps
+    {
+        [Given(@"^a regexp$")]
+        public void a_regexp() {
+        }
+    }
+}